fix(push): cap secret-bearing writes at 0600 and trim code-health hotspots (#471)
Review batch for PR #510:

- atomic_write gains max_permissions: mode preservation now intersects with
  the cap at the two secret-bearing call sites (combine_files and
  pull --merge's writer), so combined files left world-readable (0644) by
  the pre-fix write_text are tightened to 0600 on the next write instead of
  staying exposed forever; modes are only narrowed, never widened
  (review: envdrift review-bot, security)
- regression tests: cap-tightens / never-widens / default-unchanged unit
  trio (fail-proven pre-fix) + a combine_files-level tightening test
- atomic_write import moved to module level in cli_commands/sync.py
  (review: greptile P2)
- extract _build_combined_content and _encrypt_and_verify helpers, dropping
  combine_files / encrypt_secret_file / _encrypt_secrets_only_file below the
  complexity threshold (review: CodeScene complex-method x3)
- split this PR's added tests into test_push_false_success_e2e.py and
  test_partial_encryption_push_guards.py (self-contained fixtures), and
  rewrite _honest_encrypt_in_place with early-continue guards
  (review: CodeScene module-size, function-count, complex-conditional)

Fixes #471
