+++ src/envdrift/core/partial_encryption.py
@@ -386,6 +386,35 @@ def _is_fully_encrypted(file_path: Path) -> bool:
     return is_file_encrypted(file_path) and not has_plaintext_secret_value(file_path)
 
 
+def _build_combined_content(
+    env_config: PartialEncryptionEnvironmentConfig,
+    clear_lines: list[str],
+    secret_file_content: str | None,
+) -> str:
+    """Assemble the combined-file text: warning header + clear + encrypted sections.
+
+    The dotenvx public-key header block is stripped from the secret section —
+    ONLY the precise 4-line "#/" border/comment boilerplate, so a legitimate
+    user comment such as "#/ note about this key" survives into the generated
+    combined file.
+    """
+    warning = _build_warning_header(env_config.clear_file, env_config.secret_file)
+    combined_lines = warning.splitlines()
+    combined_lines.append("")
+
+    if clear_lines:
+        combined_lines.append(f"# From {env_config.clear_file}")
+        combined_lines.extend(clear_lines)
+        combined_lines.append("")
+
+    if secret_file_content:
+        stripped_text = _DOTENVX_HEADER_COMMENT_BLOCK.sub("", secret_file_content)
+        combined_lines.append(f"# From {env_config.secret_file} (encrypted)")
+        combined_lines.extend(stripped_text.splitlines())
+
+    return "\n".join(combined_lines) + "\n"
+
+
 def combine_files(
     env_config: PartialEncryptionEnvironmentConfig, *, write: bool = True
 ) -> dict[str, int | bool]:
@@ -437,31 +466,7 @@ def combine_files(
         secret_file_content = secret_file.read_text(encoding="utf-8")
         secret_lines = secret_file_content.splitlines()
 
-    # Build combined content with warning header
-    warning = _build_warning_header(env_config.clear_file, env_config.secret_file)
-
-    combined_lines = warning.splitlines()
-    combined_lines.append("")
-
-    # Add clear section
-    if clear_lines:
-        combined_lines.append(f"# From {env_config.clear_file}")
-        combined_lines.extend(clear_lines)
-        combined_lines.append("")
-
-    # Add encrypted secret section
-    if secret_lines and secret_file_content is not None:
-        # Strip ONLY the dotenvx public-key header block (the 4-line "#/"
-        # border/comment boilerplate) to avoid clutter. Matching the precise
-        # block — rather than every line starting with "#/" — preserves a
-        # legitimate user comment such as "#/ note about this key" that would
-        # otherwise be silently dropped from the generated combined file.
-        stripped_text = _DOTENVX_HEADER_COMMENT_BLOCK.sub("", secret_file_content)
-        secret_content = stripped_text.splitlines()
-        combined_lines.append(f"# From {env_config.secret_file} (encrypted)")
-        combined_lines.extend(secret_content)
-
-    new_content = "\n".join(combined_lines) + "\n"
+    new_content = _build_combined_content(env_config, clear_lines, secret_file_content)
 
     existing = combined_file.read_text(encoding="utf-8") if combined_file.exists() else None
     in_sync = existing == new_content
@@ -471,7 +476,7 @@ def combine_files(
         # same artifact holds DECRYPTED values after `pull --merge` — so write
         # it like .env.keys: 0600 for a fresh file, fchmod on the fd, atomic
         # rename. Never a bare write_text at the process umask (#471).
-        atomic_write(combined_file, new_content)
+        atomic_write(combined_file, new_content, max_permissions=0o600)
         in_sync = True
 
     # Count real secret variables (excludes comments and dotenvx public-key line)
@@ -541,6 +546,24 @@ def _git_unskip_worktree(path: Path) -> bool:
     return _run_git_update_index("--no-skip-worktree", path)
 
 
+def _encrypt_and_verify(dotenvx: DotenvxWrapper, file: Path) -> None:
+    """Encrypt ``file`` in place and verify ciphertext actually landed.
+
+    dotenvx exits 0 *without* encrypting when it cannot write or use the
+    private key (.env.keys read-only / a directory / garbage), so trusting the
+    exit code alone reports false success — re-read the file and fail loudly
+    while it is still skip-worktree protected (#471). ``DotenvxFilenameError``
+    is the pre-flight refusal of a name dotenvx would turn into an invalid key
+    (silent lockout), raised BEFORE encrypting with the plaintext intact (#467).
+    """
+    try:
+        dotenvx.encrypt(file)
+    except (DotenvxError, DotenvxFilenameError) as e:
+        raise PartialEncryptionError.encrypt_failed(file, e) from e
+    if has_plaintext_secret_value(file):
+        raise PartialEncryptionError.encryption_did_not_take_effect(file)
+
+
 def encrypt_secret_file(env_config: PartialEncryptionEnvironmentConfig) -> None:
     """
     Encrypt the secret file in-place using dotenvx.
@@ -577,24 +600,10 @@ def encrypt_secret_file(env_config: PartialEncryptionEnvironmentConfig) -> None:
         raise PartialEncryptionError.nothing_to_encrypt(secret_file)
 
     # Encrypt using dotenvx (idempotent: re-encrypts only the plaintext values,
-    # leaving any already-encrypted values untouched).
-    dotenvx = DotenvxWrapper()
-    try:
-        dotenvx.encrypt(secret_file)
-    except (DotenvxError, DotenvxFilenameError) as e:
-        # DotenvxFilenameError: the wrapper refused a name dotenvx would turn
-        # into an invalid key (silent lockout) BEFORE encrypting, so the
-        # plaintext is intact — surface a clean error, not a raw traceback (#467).
-        raise PartialEncryptionError.encrypt_failed(secret_file, e) from e
-
-    # Verify the encryption actually took effect rather than trusting the exit
-    # code: dotenvx exits 0 *without* encrypting when it cannot write or use
-    # the private key (.env.keys read-only / a directory / garbage) — the
-    # #443-era read-back the guarded encrypt backend already performs. Checked
-    # BEFORE lifting the git skip-worktree protection so a still-plaintext
-    # file stays protected from a routine `git add .` (#471).
-    if has_plaintext_secret_value(secret_file):
-        raise PartialEncryptionError.encryption_did_not_take_effect(secret_file)
+    # leaving any already-encrypted values untouched), then verify ciphertext
+    # actually landed BEFORE lifting the git skip-worktree protection so a
+    # still-plaintext file stays protected from a routine `git add .` (#471).
+    _encrypt_and_verify(DotenvxWrapper(), secret_file)
 
     # Re-enable git tracking now the file is encrypted again
     _git_unskip_worktree(secret_file)
@@ -782,18 +791,9 @@ def _encrypt_secrets_only_file(dotenvx: DotenvxWrapper, file: Path, *, check: bo
     if check:
         # Dry run: this plaintext / mixed file would be encrypted by a real push.
         return True
-    try:
-        dotenvx.encrypt(file)
-    except (DotenvxError, DotenvxFilenameError) as e:
-        # DotenvxFilenameError: a secrets_dir file whose name dotenvx can't turn
-        # into a valid key slug (e.g. "my secret.env", "café.env") is refused
-        # pre-flight with the plaintext preserved — surface it cleanly (#467).
-        raise PartialEncryptionError.encrypt_failed(file, e) from e
-    # Verify the encryption actually took effect (dotenvx exits 0 without
-    # encrypting when .env.keys is unwritable/invalid) BEFORE lifting the git
+    # Encrypt and verify ciphertext actually landed BEFORE lifting the git
     # skip-worktree protection, so a still-plaintext file stays protected (#471).
-    if has_plaintext_secret_value(file):
-        raise PartialEncryptionError.encryption_did_not_take_effect(file)
+    _encrypt_and_verify(dotenvx, file)
     # Re-enable git tracking now the file is encrypted again.
     _git_unskip_worktree(file)
     return True
